==> 01-syllabus.Rmd <==

# Syllabus {-}

**University:** European University Viadrina  
**Semester:** Winter 2025/26  
**Instructor:** Renan Magalhães (<magalhaes@kelso-institute-europe.de>)

You can download the file with the syllabus here:

**Syllabus**: **[download PDF](data/syllabus.pdf)**  

You can download the texts for the classes here. Please read them before our encounters.


FULL BOOK: Lowitzsch (2019). Energy Transition - Financing Consumer Co-Ownership.**[download PDF](data/text_general.pdf)** 

**Session 1**

Jens Lowitzsch (2019). Introduction: The Challenge of Achieving the Energy Transition. In. Energy Transitions. **[download PDF](data/text_1_1.pdf)**  

Jens Lowitzsch (2019). Conclusions: The Role of Consumer (Co-)Ownership in the Energy Transition. In. Energy Transitions. **[download PDF](data/text_1_2.pdf)**  

Renan Magalhães, Federico Narracci and Jens Lowitzsch (2025). Crowdfunding and Energy Efficiency Contracting: Exploring New Pathways for Private Investment in Building Renovations. **[download PDF](data/text_1_3.pdf)**  

Renan Magalhães, Jens Lowitzsch and Federico Narracci (2025). How (Co-)Ownership in Renewables Improves Heating Usage Behaviour and the Willingness to Adopt Energy-Efficient Technologies—Data from German Households. **[download PDF](data/text_1_4.pdf)**  


**Session 2**

Florian Hanke and Jens Lowitzsch (2021). Empowering Vulnerable Consumers to Join Renewable Energy Communities—Towards an Inclusive Design of the Clean Energy Package. **[download PDF](data/text_2_1.pdf)**

Jens Lowitzsch, Monika Bucha and Sarah Lonscher (2025). From Access to Ownership Energy Communities & Social Inclusion in the EU’s Energy Transition. **[download PDF](data/text_2_2.pdf)**

Christina E. Hoicka, Jens Lowitzsch, Marie Claire Brisbois, Ankit Kumar and Luis Ramirez Camargo (2021). Implementing a just renewable energy transition: Policy advice for transposing the new European rules for renewable energy communities. **[download PDF](data/text_2_3.pdf)**


## WATCH: Electricity System: an overview
<iframe src="https://drive.google.com/file/d/14b2cZH0DO4aa2UYZPSXXXOMKhVFkvz6H/preview"
        width="640" height="360" allow="autoplay"></iframe>

THIS LINK >>>> https://miro.com/welcomeonboard/djFRTGV0OHBnUzZ6QnRKOEl5OG51aHZhTU16ZmVvdWRLdHBuOTJ6Ulg5STJxTGlJYkhSajFYc1FPOSt6RDZMMmlTTlJ2SUNVWWtEa0QrVVJUTnlKSFJaUjlaMnJLWTZUQi9TQjEwaVRWQnR6U2Yxb1V5dWRNaXpaUUk0QVh4bXhQdGo1ZEV3bUdPQWRZUHQzSGl6V2NBPT0hdjE=?share_link_id=830981808939
